Refactor Component Code
