+++ src/components/Page.js
@@ -1,0 +1,18 @@
+import styled from "styled-components";
+
+export default styled.div`
+  @import url("https://fonts.googleapis.com/css2?family=Raleway:wght@586&display=swap");
+  font-family: "Raleway", sans-serif;
+  color: #003;
+
+  > h1 {
+    margin-bottom: 0;
+  }
+  > p {
+    display: flex;
+    flex-direction: row;
+    justify-content: flex-start;
+    align-items: center;
+    margin-top: 0;
+  }
+`;
