+++ src/lib/index.js
@@ -1,3 +1,8 @@
+export * from "./totalTime";
+export * from "./icons";
+export * from "./toTree";
+export * from "./TimeDisplay";
+
 export const toJSON = (res) => res.json();
 export const toText = (res) => res.text();
