+++ src/lib/totalTime.js
